I started by adding a landscape class and wrote some general descriptions of weather and landscape class.
I deleted the .cpp files as I will first define some general function and classes.
This is supposed to first create a general structure so that
I can later work within specified confines for each function/class and
no longer have to worry about what and what not it's supposed to do.